update import & export refacto
