+++ src/components/import.js
@@ -4,3 +4,8 @@ console.log(add(16, 4));
 
 // const add = require("./calculate").add
 
+
+
+
+// console.log(typeof randomSold);
+
